+++ CHANGELOG.md
@@ -15,4 +15,7 @@ Note: The format of the versioning is **based [Semantic Versioning 2.0.0](https:
 ## Stories
 Format: version number, story identifier, date (mm/dd/yyy), story title
 
+- [0.4.0] ``AC2`` [01/15/2020] Create Palette List using Mock Values
+- [0.3.0] ``AC1`` [01/15/2020] Create Site Pallete
 - [0.2.0] ``AC1`` [01/15/2020] Add Font Awesome
+- [0.1.0] ``master`` [01/15/2020] Base of Project